requirements.txt
# 📁 requirements.txt -----

Authlib==1.3.0
boto3==1.34.13
Flask==2.2.5
matplotlib==3.8.0
mysql-connector-python==8.1.0
numpy==1.26.0
opencv-python==4.8.1.78
pandas==2.1.1
python-dotenv==1.0.0
scikit-fuzzy==0.4.2
SQLAlchemy==2.0.21
urllib3==1.25.10
uuid==1.30
datetime==5.4

# pip install git+https://github.com/facebookresearch/segment-anything.git
# !pip install torch torchvision opencv-python numpy